requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
langchain-google-genai==2.0.10
langgraph==0.2.55
google-generativeai==0.8.3
gspread==6.1.2
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
slack-sdk==3.33.0
notion-client==2.2.1
hubspot-api-client==9.0.0
tenacity==8.2.3
google-api-python-client==2.139.0